requirements.txt
opencv-python<4.9
numpy<2
Pillow
scikit-learn
torch
torchvision